landing: Add support for note pushes!
Successfully completed a note push. Currently it goes out to all devices
as I need to retrieve the device iden from the nickname to push to a
specific device. Additionally, the title doesn't work for the note push.
Lastly, somewhere along the line I need to register the device as a
separate device. Currently, the pushes say from "metroPushv1".